feat: enhance ONVIF recording, Dahua encoding detection, and Xiaomi HLS forwarding
- Add ONVIF recording detail endpoint and service availability checks
- Improve ONVIF namespace parsing for Media1/Media2 service detection
- Add Dahua camera encoding detection via VideoEncoderConfigurations
- Add Xiaomi recorder HLS forwarding and audio broadcast via StreamHub
- Fix SOAP WSSE timestamp to use UTC
- Add SSE auth token support for AI events and ONVIF events
- Handle duplicate in-stream errors gracefully in camera manager
- Fix Quality field type to float64 for ONVIF media config
- Update protocols test to include new protocol types
- Improve ONVIF recording query UI with drill-down detail view
- Truncate long profile names in camera cards and live view

+++ onvif/media_config.go
@@ -147,7 +147,7 @@ func (s *MediaService) GetVideoEncoderConfigurations(ctx context.Context) ([]Vid
 				Width  int `xml:"Width"`
 				Height int `xml:"Height"`
 			} `xml:"Resolution"`
-			Quality      int `xml:"Quality"`
+			Quality      float64 `xml:"Quality"`
 			RateControl  struct {
 				FrameRateLimit int `xml:"FrameRateLimit"`
 				BitrateLimit   int `xml:"BitrateLimit"`
@@ -170,7 +170,7 @@ func (s *MediaService) GetVideoEncoderConfigurations(ctx context.Context) ([]Vid
 			UseCount:       c.UseCount,
 			Encoding:       c.Encoding,
 			Resolution:     Resolution{Width: c.Resolution.Width, Height: c.Resolution.Height},
-			Quality:        c.Quality,
+			Quality:        int(c.Quality),
 			FramerateLimit: c.RateControl.FrameRateLimit,
 			BitrateLimit:   c.RateControl.BitrateLimit,
 			GovLength:      c.H264.GovLength,
